Implement custom '[chill route ...]' shortcodes
Uses a jinja2 template filter 'shortcodes' to parse the content for
[chill route ...] shortcodes.  The '...' here is a url route
( /path/to/something ) that is rendered the same way that a page at that
url is.  This allows insertion of content.  Main use case is custom HTML
for inserting media within user content like adding a <picture> and
such.