BREAKING CHANGE: Check query method for db commits
Only mutable http request methods (POST, PUT, PATCH, DELETE) are able to
make changes to the database. A GET request will no longer be able to
commit any database changes.
The 'method' value can no longer be passed in as a value to the mutable
request methods as well. It was replaced with the actual request method
anyways.